# Conflicts:
#	backend/src/main/java/com/socialnetwork/socialnetworkapi/service/DefaultNotificationService.java
#	frontend/src/api/messages.js
#	frontend/src/components/Header/Navigation/Navigation.jsx
#	frontend/src/components/Modal/Popup/PopupLogOut.jsx
#	frontend/src/components/Sidebar/Sidebar.jsx
#	frontend/src/page/Bookmarks/Bookmarks.jsx
#	frontend/src/page/Bookmarks/bookmarks.style.scss
#	frontend/src/page/Bookmarks/noBookmarks.jsx
#	frontend/src/redux/tokenSlice.js

+++ backend/src/main/java/com/socialnetwork/socialnetworkapi/service/DefaultCommunityService.java
@@ -1,10 +1,6 @@
 package com.socialnetwork.socialnetworkapi.service;
 
-import com.socialnetwork.socialnetworkapi.dao.repository.CommunityMembersRepository;
-import com.socialnetwork.socialnetworkapi.dao.repository.CommunityRepository;
-import com.socialnetwork.socialnetworkapi.dao.repository.FavoritesRepository;
-import com.socialnetwork.socialnetworkapi.dao.repository.LikesRepository;
-import com.socialnetwork.socialnetworkapi.dao.repository.PostRepository;
+import com.socialnetwork.socialnetworkapi.dao.repository.*;
 import com.socialnetwork.socialnetworkapi.dao.service.CommunityService;
 import com.socialnetwork.socialnetworkapi.dto.community.CommunityCreateRequest;
 import com.socialnetwork.socialnetworkapi.dto.community.CommunityRequest;
@@ -20,11 +16,14 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.UUID;
 
 @Service
+@Transactional
 @RequiredArgsConstructor
 public class DefaultCommunityService implements CommunityService {
 
@@ -38,9 +37,9 @@ public class DefaultCommunityService implements CommunityService {
     @Override
     public CommunityResponse createCommunity(CommunityCreateRequest req) {
         if (!communityRepository.existsByName(req.getName())) {
-            CommunityResponse resp = mapper.communityToDTO(communityRepository.save(mapper.communityFromDTO(req)));
-            communityMembersRepository.save(new CommunityMember(req.getCreatorId(), resp.getId(), CommunityRole.ADMINISTRATOR.name()));
-            return resp;
+            UUID communityId = communityRepository.save(mapper.communityFromDTO(req)).getId();
+            communityMembersRepository.save(new CommunityMember(req.getCreatorId(), communityId, CommunityRole.ADMINISTRATOR.name()));
+            return this.toDtoCurrentUserFetched(communityId, req.getCreatorId());
         } else {
             throw new BadRequestException("Community with name " + req.getName() + " already exists");
         }
@@ -48,30 +47,41 @@ public class DefaultCommunityService implements CommunityService {
 
 
     @Override
-    public Boolean deleteCommunity(UUID req) {
-        try {
-            postRepository.getAllByCommunityId(req).forEach(post -> {
-                        likesRepository.deleteAllByPostId(post.getId());
-                        favoritesRepository.deleteAllByPostId(post.getId());
-                    }
-            );
-            postRepository.deleteAllByCommunityId(req);
-            communityMembersRepository.deleteAllByCommunityId(req);
-            communityRepository.deleteById(req);
-            return true;
-        } catch (Exception e) {
-            return false;
+    public Boolean deleteCommunity(UUID requestedCommId, UUID userId) {
+        if (Objects.equals(communityMembersRepository.getByCommunityIdAndUserId(requestedCommId, userId).getRole(), CommunityRole.ADMINISTRATOR.toString())) {
+            System.out.println("if passed successfully,");
+            try {
+                postRepository.getAllByCommunityId(requestedCommId).forEach(post -> {
+                            likesRepository.deleteAllByPostId(post.getId());
+                            favoritesRepository.deleteAllByPostId(post.getId());
+                        }
+                );
+                postRepository.deleteAllByCommunityId(requestedCommId);
+                communityMembersRepository.deleteAllByCommunityId(requestedCommId);
+                communityRepository.deleteById(requestedCommId);
+                return true;
+            } catch (Exception e) {
+                System.out.println(e);
+                return false;
+            }
         }
+        return false;
     }
 
     @Override
-    public CommunityResponse getById(UUID req) {
-        return this.toDtoFetched(req);
+    public CommunityResponse getById(UUID req, UUID currentUserId) {
+        return this.toDtoCurrentUserFetched(req, currentUserId);
     }
 
     private CommunityResponse toDtoFetched(UUID req) {
         CommunityResponse resp = mapper.communityToDTO(communityRepository.findById(req).orElseThrow());
         resp.setMembersCounts(communityMembersRepository.countAllByCommunityId(req));
+        return resp;
+    }
+
+    private CommunityResponse toDtoCurrentUserFetched(UUID req, UUID currentUserId) {
+        CommunityResponse resp = this.toDtoFetched(req);
+        resp.setFollowed(communityMembersRepository.existsByUserIdAndCommunityId(currentUserId, req));
         return resp;
     }
 
@@ -83,7 +93,7 @@ public class DefaultCommunityService implements CommunityService {
     @Override
     public List<CommunityResponse> getPaged(CommunityRequest req) {
         Pageable pageable = PageRequest.of(req.getPage(), req.getPageSize(), Sort.by("createdAt").descending());
-        return communityRepository.findAll(pageable).stream().map(community -> this.toDtoFetched(community.getId())).toList();
+        return communityRepository.findAll(pageable).stream().map(community -> this.toDtoCurrentUserFetched(community.getId(), req.getUserId())).toList();
     }
 
     @Override
@@ -96,6 +106,12 @@ public class DefaultCommunityService implements CommunityService {
         communityMembersRepository.save(mapper.communityMemberFromDTO(req));
         return true;
 
+    }
+    @Override
+    public List<CommunityResponse> getAllByMemberId(UUID userId, Integer page){
+        return page == null
+                ? communityMembersRepository.getAllByUserId(userId).stream().map(communityMember -> this.toDtoCurrentUserFetched(communityMember.getCommunityId(), userId)).toList()
+                : communityMembersRepository.findAllByUserId(userId, PageRequest.of(page, 10, Sort.by("createdAt").descending())).stream().map(communityMember -> this.toDtoCurrentUserFetched(communityMember.getCommunityId(), userId)).toList();
     }
 
     @Override
